% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/childmortality.R
\name{childhoodmortality}
\alias{childhoodmortality}
\title{Calculates childhood mortality}
\usage{
childhoodmortality(data, grouping)
}
\arguments{
\item{data}{dataframe object containing DHS variables YEAR, PSU, PERWEIGHT, KIDDOBCMC, INTDATECMC, KIDAGEDIEDIMP, and grouping var.}

\item{grouping}{This character string gives the variable name of the variable denoting groups.}
}
\description{
calculates childhood mortality & SE's
}
