% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/childhoodmortality.R
\name{childhoodmortality}
\alias{childhoodmortality}
\title{Calculates childhood mortality}
\usage{
childhoodmortality(data, grouping, rate_type = "underfive")
}
\arguments{
\item{data}{dataframe object containing DHS variables YEAR, PSU, PERWEIGHT, KIDDOBCMC, INTDATECMC, KIDAGEDIEDIMP, and grouping var.}

\item{grouping}{This character string gives the variable name of the variable denoting groups.}

\item{rate_type}{This character string gives the type of mortality rate to be calculated (neonatal, postneonatal, infant, child, under-five)}
}
\description{
calculates childhood mortality & SE's
}
\examples{
data("model_ipums_dhs_dataset")
underfive_mortality_rates <- childhoodmortality(
 model_ipums_dhs_dataset,
 grouping ="wealthq",
 rate_type = "underfive"
)

}
